+++ playgrounds/desktop/src/table/index.vue
@@ -1,12 +1,5 @@
 <template>
-  <u-table
-    style="height: 500px"
-    :data="students"
-    :columns="columns"
-    v-model:checked="checked"
-    show-index
-    checkable
-  >
+  <u-table style="height: 500px" :data="students" :columns="columns" v-model:checked="checked">
     <template #column:action>
       <u-action-group :max="4">
         <u-action need-confirm type="danger">删除</u-action>
@@ -35,11 +28,11 @@ const students = Array.from({ length: 200 }).map((_, i) => {
 const checked = shallowRef([])
 
 const columns = defineTableColumns([
-  { key: 'name', name: '姓名' },
-  { key: 'age', name: '年龄', align: 'center' },
-  { key: 'grade', name: '年级', align: 'center' },
-  { key: 'class', name: '班级', align: 'center' },
-  { key: 'score', name: '分数', align: 'center' },
-  { key: 'action', name: '操作', align: 'center' }
+  { key: 'name', name: '姓名' }
+  // { key: 'age', name: '年龄', align: 'center' },
+  // { key: 'grade', name: '年级', align: 'center' },
+  // { key: 'class', name: '班级', align: 'center' },
+  // { key: 'score', name: '分数', align: 'center' },
+  // { key: 'action', name: '操作', align: 'center' }
 ])
 </script>