+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @veltra/icons
 
+## 1.0.9
+
+### Patch Changes
+
+- 81dbe41: 重构 use-lock -> use-user-action, 更加符合直觉
+- 组件优化
+
 ## 1.0.8
 
 ### Patch Changes